backend/requirements.txt
asgiref==3.3.1
celery==5.0.5
Django==2.2.16
django-cors-headers==3.7.0
django_celery_beat==2.2.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.13
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-extensions==0.7.0
jsonpath==0.82
mongoengine==0.22.1
pymysql==0.10.0
Pillow==8.1.0
PyJWT==1.7.1
pymongo==3.11.3
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
xlwt==1.3.0
xlrd==2.0.1
coreapi==2.3.3
user-agents==2.2.0
eventlet==0.30.2
psutil==5.8.0
drf-yasg==1.20.0
daphne==3.0.2
channels==3.0.3
channels-redis==3.2.0
mitmproxy~=6.0.2

opencv-python~=4.4.0.42
enum34~=1.1.10
hpack~=4.0.0
hyperframe~=6.0.0
h2~=4.0.0
dnspython~=1.16.0
idna~=2.10
future~=0.18.2
pytest~=5.4.3
h11~=0.12.0
cryptography~=3.4.7
typing~=3.7.4.3
cffi~=1.14.6
numpy~=1.19.1
pip~=21.3.1
colorlog~=6.5.0
toml~=0.10.2
Brotli~=1.0.9
retry~=0.9.2
py~=1.10.0
lxml~=4.6.3
tornado~=6.1
amqp~=5.0.6
vine~=5.0.0
attrs~=20.3.0
colorama~=0.4.4
Jinja2~=2.11.2
Werkzeug~=2.0.1
itsdangerous~=2.0.1
click~=7.1.2
blinker~=1.4
kombu~=5.2.2
billiard~=3.6.4.0
SQLAlchemy~=1.4.22
zstandard~=0.14.1
pyasn1~=0.4.8
hiredis~=1.1.0
Twisted~=21.7.0
txaio~=21.2.1
urwid~=2.1.2
Pygments~=2.7.4
greenlet~=1.1.0
gevent~=21.1.2
autobahn~=21.3.1
PyYAML~=5.4.1
MarkupSafe~=0.23
pyparsing~=2.4.7
wsproto~=1.0.0
Automat~=20.2.0
certifi~=2020.6.20
itypes~=1.2.0
coreschema~=0.0.4
requests~=2.24.0
uritemplate~=4.1.1
passlib~=1.7.4
monotonic~=1.5
incremental~=21.3.0
constantly~=15.1.0
bcrypt~=3.2.0
hyperlink~=21.0.0
appdirs~=1.4.4
urllib3~=1.25.10
wcwidth~=0.2.5
aioredis~=1.3.1
msgpack~=1.0.2
msgpack-python~=0.4.7
python-dateutil~=2.8.1
inflection~=0.5.1
chardet~=3.0.4
kaitaistruct~=0.9
publicsuffix2~=2.20191221
pyperclip~=1.8.1
sortedcontainers~=2.3.0
ldap3~=2.8.1
Flask~=1.1.2
pycparser~=2.20
setuptools~=56.0.0
python-crontab~=2.6.0
Markdown~=3.3.4
django-timezone-field~=4.2.1
